Moved the two step transfer into a separate library and splitted up tests
