+++ tests/mocks/Ownable.cairo
@@ -26,16 +26,6 @@ func owner{
     return (owner=owner)
 end
 
-@view
-func proposedOwner{
-        syscall_ptr : felt*,
-        pedersen_ptr : HashBuiltin*,
-        range_check_ptr
-    }() -> (proposed_owner: felt):
-    let (proposed_owner) = Ownable.proposed_owner()
-    return (proposed_owner=proposed_owner)
-end
-
 @external
 func transferOwnership{
         syscall_ptr : felt*,
@@ -43,36 +33,6 @@ func transferOwnership{
         range_check_ptr
     }(new_owner: felt):
     Ownable.transfer_ownership(new_owner)
-    return ()
-end
-
-@external
-func proposeOwner{
-        syscall_ptr : felt*,
-        pedersen_ptr : HashBuiltin*,
-        range_check_ptr
-    }(proposed_owner: felt):
-    Ownable.propose_owner(proposed_owner)
-    return ()
-end
-
-@external
-func acceptOwnership{
-        syscall_ptr : felt*,
-        pedersen_ptr : HashBuiltin*,
-        range_check_ptr
-    }():
-    Ownable.accept_ownership()
-    return ()
-end
-
-@external
-func cancelOwnershipProposal{
-        syscall_ptr : felt*,
-        pedersen_ptr : HashBuiltin*,
-        range_check_ptr
-    }():
-    Ownable.cancel_ownership_proposal()
     return ()
 end
